requirements.txt
# Core dependencies
langgraph>=0.2.0
dspy-ai>=2.4.0
langsmith>=0.1.0
openai>=1.12.0
google-generativeai>=0.3.0

# Data processing
pydantic>=2.5.0
pyyaml>=6.0.1
pandas>=2.1.0
numpy>=1.26.0,<2.0.0  # Pin to NumPy 1.x to avoid compatibility issues

# NLP and quality metrics
# Note: Using Gemini embeddings instead of sentence-transformers
# If Gemini is unavailable, code will fall back to word-overlap based similarity.
nltk>=3.8.1
textstat>=0.7.3  # Optional - has fallback

# API and CLI
fastapi>=0.109.0
uvicorn>=0.27.0
click>=8.1.7
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0  # Required for loading .env file
tqdm>=4.66.0
tenacity>=8.2.3
